Push boilerplate `Logger` implementation up to `BuildOperationAwareLogger`
To make it easier to introduce new subclasses.
